fix: use uname -r to fill KERNEL_VER
Using a linux asahi, if I use `uname -r` to fill the value
of KERNEL_VER it work like a charm, else it retrive a wrong
version as the kernel is using a patched version for 16k
page size.

Using the current method in this case evoke a wrong directory
because the 'source' directory reference a not existing path.

BTW: I can override the value by:

KERNEL_VER="$(uname -r)" make

But wondering if `uname -r` would fit all the cases.

Signed-off-by: Alejandro Visiedo <alejandro.visiedo@gmail.com>